Splitte opp utils-filer
