+++ web/src/frontend/src/nav-soknad/utils/animationUtils.ts
@@ -1,0 +1,9 @@
+import { scroller } from "react-scroll";
+
+export function scrollToElement(id: string) {
+	scroller.scrollTo(id, {
+		duration: 500,
+		smooth: "easeOutQuad",
+		offset: -30
+	});
+}
